requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
boto3==1.34.55
    # via -r requirements.in
botocore==1.34.55
    # via
    #   boto3
    #   s3transfer
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.6
    # via anyio
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pydantic==2.6.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
s3transfer==0.10.0
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.36.3
    # via fastapi
typing-extensions==4.10.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
urllib3==2.0.7
    # via botocore
uvicorn==0.27.1
    # via -r requirements.in
